requirements.txt
einops
ftfy
h5py
pytorch-lightning==1.9.5
regex
scikit-video
tqdm
imageio-ffmpeg
moviepy
av
omegaconf==2.1.1
pandas
matplotlib
seaborn
scikit-learn
basemap
basemap-data
basemap-data-hires
easydict
imageio
netCDF4
moviepy
imageio-ffmpeg
opencv-python
torchmetrics
torchvision

# additional dependencies required by SAL and PSD utilities
pysteps
scipy
joblib
pyfftw
dask
toolz
